henson-backend/requirements/prod.txt
Flask==3.1.1
Werkzeug==3.1.3
flask-cors
supabase
python-dotenv
gunicorn